refactor : rename ChatRoomService
